Respond to feedback.

+++ enterprise/internal/codeintel/store/mocks/mock_store.go
@@ -197,7 +197,7 @@ func NewMockStore() *MockStore {
 			},
 		},
 		DequeueFunc: &StoreDequeueFunc{
-			defaultHook: func(context.Context, int) (store.Upload, store.Store, bool, error) {
+			defaultHook: func(context.Context, int64) (store.Upload, store.Store, bool, error) {
 				return store.Upload{}, nil, false, nil
 			},
 		},
@@ -1312,15 +1312,15 @@ func (c StoreDeleteUploadsWithoutRepositoryFuncCall) Results() []interface{} {
 // StoreDequeueFunc describes the behavior when the Dequeue method of the
 // parent MockStore instance is invoked.
 type StoreDequeueFunc struct {
-	defaultHook func(context.Context, int) (store.Upload, store.Store, bool, error)
-	hooks       []func(context.Context, int) (store.Upload, store.Store, bool, error)
+	defaultHook func(context.Context, int64) (store.Upload, store.Store, bool, error)
+	hooks       []func(context.Context, int64) (store.Upload, store.Store, bool, error)
 	history     []StoreDequeueFuncCall
 	mutex       sync.Mutex
 }
 
 // Dequeue delegates to the next hook function in the queue and stores the
 // parameter and result values of this invocation.
-func (m *MockStore) Dequeue(v0 context.Context, v1 int) (store.Upload, store.Store, bool, error) {
+func (m *MockStore) Dequeue(v0 context.Context, v1 int64) (store.Upload, store.Store, bool, error) {
 	r0, r1, r2, r3 := m.DequeueFunc.nextHook()(v0, v1)
 	m.DequeueFunc.appendCall(StoreDequeueFuncCall{v0, v1, r0, r1, r2, r3})
 	return r0, r1, r2, r3
@@ -1328,7 +1328,7 @@ func (m *MockStore) Dequeue(v0 context.Context, v1 int) (store.Upload, store.Sto
 
 // SetDefaultHook sets function that is called when the Dequeue method of
 // the parent MockStore instance is invoked and the hook queue is empty.
-func (f *StoreDequeueFunc) SetDefaultHook(hook func(context.Context, int) (store.Upload, store.Store, bool, error)) {
+func (f *StoreDequeueFunc) SetDefaultHook(hook func(context.Context, int64) (store.Upload, store.Store, bool, error)) {
 	f.defaultHook = hook
 }
 
@@ -1336,7 +1336,7 @@ func (f *StoreDequeueFunc) SetDefaultHook(hook func(context.Context, int) (store
 // Dequeue method of the parent MockStore instance inovkes the hook at the
 // front of the queue and discards it. After the queue is empty, the default
 // hook function is invoked for any future action.
-func (f *StoreDequeueFunc) PushHook(hook func(context.Context, int) (store.Upload, store.Store, bool, error)) {
+func (f *StoreDequeueFunc) PushHook(hook func(context.Context, int64) (store.Upload, store.Store, bool, error)) {
 	f.mutex.Lock()
 	f.hooks = append(f.hooks, hook)
 	f.mutex.Unlock()
@@ -1345,7 +1345,7 @@ func (f *StoreDequeueFunc) PushHook(hook func(context.Context, int) (store.Uploa
 // SetDefaultReturn calls SetDefaultDefaultHook with a function that returns
 // the given values.
 func (f *StoreDequeueFunc) SetDefaultReturn(r0 store.Upload, r1 store.Store, r2 bool, r3 error) {
-	f.SetDefaultHook(func(context.Context, int) (store.Upload, store.Store, bool, error) {
+	f.SetDefaultHook(func(context.Context, int64) (store.Upload, store.Store, bool, error) {
 		return r0, r1, r2, r3
 	})
 }
@@ -1353,12 +1353,12 @@ func (f *StoreDequeueFunc) SetDefaultReturn(r0 store.Upload, r1 store.Store, r2 
 // PushReturn calls PushDefaultHook with a function that returns the given
 // values.
 func (f *StoreDequeueFunc) PushReturn(r0 store.Upload, r1 store.Store, r2 bool, r3 error) {
-	f.PushHook(func(context.Context, int) (store.Upload, store.Store, bool, error) {
+	f.PushHook(func(context.Context, int64) (store.Upload, store.Store, bool, error) {
 		return r0, r1, r2, r3
 	})
 }
 
-func (f *StoreDequeueFunc) nextHook() func(context.Context, int) (store.Upload, store.Store, bool, error) {
+func (f *StoreDequeueFunc) nextHook() func(context.Context, int64) (store.Upload, store.Store, bool, error) {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
@@ -1396,7 +1396,7 @@ type StoreDequeueFuncCall struct {
 	Arg0 context.Context
 	// Arg1 is the value of the 2nd argument passed to this method
 	// invocation.
-	Arg1 int
+	Arg1 int64
 	// Result0 is the value of the 1st result returned from this method
 	// invocation.
 	Result0 store.Upload
